auth/requirements.txt
Flask==2.1.2
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-jwt-extended==4.4.2
flask_authorize==0.2.6
flask-mail==0.9.1
flasgger==0.9.5
gunicorn==20.1.0
psycopg2-binary==2.9.3
sqlalchemy==1.4.39
flask_jwt_extended==4.4.2
flask_authorize==0.2.6
flasgger==0.9.5
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
alembic==1.8.1
redis==4.1
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0
requests==2.28.1
requests-oauthlib==1.3.1
